experiments/imc-01-agentic-bullwhip-effect/requirements.txt
openai>=1.12.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
jupyter>=1.0.0
